k8s-processor/requirements.txt
azure-storage-blob
azure-servicebus
pillow
